Explicitly convert enum to int for QString::arg
The current code doesn't compile after https://invent.kde.org/qt/qt/qtbase/-/commit/0a1c69bba5b5dd26b937ac0a67ab0bdd1c32b221

+++ src/kar.cpp
@@ -84,7 +84,7 @@ bool KAr::openArchive(QIODevice::OpenMode mode)
         return true;
     }
     if (mode != QIODevice::ReadOnly && mode != QIODevice::ReadWrite) {
-        setErrorString(tr("Unsupported mode %1").arg(mode));
+        setErrorString(tr("Unsupported mode %1").arg(static_cast<int>(mode)));
         return false;
     }
